choosing page

+++ game/skills.py
@@ -4,8 +4,8 @@ from dataclasses import dataclass
 @dataclass
 class Skill:
     name: str
-    damage: int
-    stamina: int
+    damage: float
+    stamina: float
 
 
 ferocious_kick = Skill(name='Свирепый пинок', damage=22, stamina=6)
